+++ package.json
@@ -13,8 +13,11 @@
     "react-datepicker": "^4.21.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.11.0",
+    "react-redux": "^9.1.0",
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
+    "redux": "^5.0.1",
+    "redux-thunk": "^3.1.0",
     "uuid": "^9.0.0",
     "web-vitals": "^2.1.4"
   },
